requirements.txt
gunicorn
numpy
Flask
pandas
transformers
faiss-cpu